refactor data parser
