+++ data/preprocessing.cpp
@@ -1,0 +1,10 @@
+#include <iostream>
+#include <fstream>
+#include <sstream>
+
+using namespace std;
+
+array<float> normalize_data(input_data){
+
+  
+}
