--- data/read_file.h
@@ -1,26 +1,0 @@
-// COPYRIGHTS AND PERMISSIONS:
-// Copyright 2021 MORSE Corp All rights reserved.
-//
-// Subject to any rights the U.S. Government may have in the Software,
-// permission is granted to the person rightfully obtaining a copy of this
-// Software from MORSE Corp to use, copy, modify and distribute the
-// Software only as expressly authorized in writing from MORSE Corp.
-//
-// The above copyright notice and this permission notice shall be included
-// in all permitted copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
-// OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
-// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
-// IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
-// CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
-// TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
-// SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-/// @file
-#ifndef SVM_READ_FILE_H
-#define SVM_READ_FILE_H
-
-
-void read_csv();
-
-#endif // SVM_READ_FILE_H
